+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @enschedule/cli
 
+## 1.1.5
+
+### Patch Changes
+
+- Make description optional
+- Updated dependencies
+  - @enschedule/types@1.1.4
+  - @enschedule/worker-api@1.1.4
+
 ## 1.1.4
 
 ### Patch Changes