+++ packages/worker/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @enschedule/worker
 
+## 1.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @enschedule/pg-driver@1.1.5
+  - @enschedule/types@1.1.4
+  - @enschedule/worker-api@1.1.4
+
 ## 1.1.4
 
 ### Patch Changes